Add files via upload
This is  an application which will display the date (CSV date field)
corresponding to the maximum value (export or import depending on what the user asks).
Suppose the user types $ trade exports China, my program will display the date on which New Zealand had the largest exports to China. If one type $ trade imports China the program will display the date on which New Zealand had the largest imports from China.